+++ NEWS.md
@@ -1,6 +1,8 @@
 # TreeSearch 1.6.1.9004 (development)
 - Support single-character matrices in `ClusteringConcordance()`
 - Fix `DoNothing(x)` to return `x` (not `NULL`)
+- `PresCont()` implements the Group Present / Contradicted measure of
+  Goloboff et al. (2003).
 - `Consistency()` also returns the relative homoplasy index of Steell et al. 
   (2023).
 - Remove unused `delete_rawdata()` due to implementation issues.